lock upx version to v4.0.1
submodules no longer in git

I removed .git files for all subrepos. So all subrepo files are in the
main repo. Did this because of troubles to include the correct
submodule versions in yocto.

These are the "expanded" submodules are using v4.0.1-upx-vendor branch:
- doctest
- lzma-sdk
- ucl
- valgrind
- zlib

There is one exception with valgrind which I think uses
v4.0.2-upx-vendor instead.